+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Hello, Text!
 - Support unbordered `text` widget if SWT style `:none` was passed
 - Support `:center`, `:left`, `:right` SWT styles for `text` widget
+- Support `:read_only` SWT style for `text` widget
 
 ## 0.26.3